add 'force_pass' to TestContext
This commit introduces a 'force_pass' variable to TestContext and
handles it in the test runner. The idea is for tests to be able to
refuse to run if their parametrisation does not match the environemnt.

'force_pass' is meant to be used dinamically within the test
(i.e. the test starts and realises that it shouldn't be running
and throws after setting 'force_pass' on the context).